sm8550-common: SEPolicy: remove remnants from old tree
We do not use SettingsLib in this tree, it was used in our old one, it is not defined so remove it from SEPolicy in order to be able to build.

sm8550-common: SEPolicy: remove remnants of hal_misys from old tree

files for hal_misys do not exist in this tree and we do not use it. It has been left from old tree sepolicies.

sm8550-common: SEPolicy: do not allow cameraserver to get vendor_default_prop

this triggers a neverallow rule, so let's remove it. Camera works without it anyway. Just triggers a AVC denial in the log.

sm8550-common: SEPolicy: define types for missing devices

- type stmvl53l5_device, dev_type;
- type camlog_device, dev_type;
- type migt_dev, dev_type, mlstrustedobject;

sm8550-common: SEPolicy: Remove Neverallows from platform_app

Both of these trigger neverallows and the camera works without them, it will just trigger an avc denial in the logs. We will remove them.

sm8550-common: SEPolicy: remove remnants of DualCameraCal

we do not use DualCameraCal in this tree, it was used in our old one, it is not defined so remove it from SEPolicy in order to be able to build.